+++ src/WithSettingsInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Avtocod\B2BApi;
+
+interface WithSettingsInterface
+{
+    /**
+     * Get client settings object.
+     *
+     * @return Settings
+     */
+    public function getSettings(): Settings;
+}